+++ pytorch_lightning/trainer/connectors/callback_connector.py
@@ -94,7 +94,7 @@ class CallbackConnector:
                 " bar pass `enable_progress_bar = False` to the Trainer."
             )
 
-        self.configure_progress_bar(progress_bar_refresh_rate, process_position, enable_progress_bar)
+        self._configure_progress_bar(progress_bar_refresh_rate, process_position, enable_progress_bar)
 
         # configure the ModelSummary callback
         self._configure_model_summary_callback(enable_model_summary, weights_summary)
@@ -211,7 +211,7 @@ class CallbackConnector:
         if not existing_swa:
             self.trainer.callbacks = [StochasticWeightAveraging()] + self.trainer.callbacks
 
-    def configure_progress_bar(
+    def _configure_progress_bar(
         self, refresh_rate: Optional[int] = None, process_position: int = 0, enable_progress_bar: bool = True
     ) -> None:
         progress_bars = [c for c in self.trainer.callbacks if isinstance(c, ProgressBarBase)]
@@ -238,8 +238,7 @@ class CallbackConnector:
         if refresh_rate == 0 or not enable_progress_bar:
             return
         if refresh_rate is None:
-            # smaller refresh rate on colab causes crashes, choose a higher value
-            refresh_rate = 20 if os.getenv("COLAB_GPU") else 1
+            refresh_rate = 1
 
         progress_bar_callback = TQDMProgressBar(refresh_rate=refresh_rate, process_position=process_position)
         self.trainer.callbacks.append(progress_bar_callback)